Extract saveBalance and loadBalance from modules to UnicacityAddon
